feat: Add "Favorites" section and "See All" screen
Implemented the UI to prominently display favorited (fixed) tasks. The HomeScreen now features a horizontal list for favorited tasks. A new SeeAllScreen was added to show a full list of these tasks, with navigation enabled from the "See All" button. Also corrected the navigation path for the "Create Note" button in the NavBar.